Small logo fix, killed a few sections, added portfolio i frames
READ ME!!
So I have this idea that might be awesome in theory but end up being
garbage in practice.
Check out the the portfolio section. You have LIVE embeds of our sites
on mockups. Cool AF right?? It’s causing an issue where it pushes down
the descriptions that we’ll need to address.
Give me your opinion if it’s worth the load time or if we should stick
to static images.
\m/